Add support for xattr synchronization
Support extended attributes synchronization as file properties. Since
all properties are stored in the archive, this means that all xattrs
will be stored verbatim in the archive. To not break the archive, xattr
values are currently limited to 383 bytes (but the number of attributes
itself is not limited). Many platforms allow very large or even
arbitrarily long xattr values; it is not feasible to synchronize such
values as Props.

Add new preference to request/enable xattr synchronization.

Props.get interface is changed to include the path and name, which
is needed to retrieve the xattrs.

Archive and wire formats change.