Use ctime to fast metadata updates checking
mtime is already used for fast update checking. If mtime has not changed
(and some other conditions hold true) then the file has not changed and
its contents do not have to be scanned for updates.

Until now, this fast update checking has not applied to file metadata in
the same way. It has not been needed because all the synced properties
were available from 'stat' which was called anyway to get the mtime used
in fast update checking.

This is changing with extended metadata such as ACL and xattrs. Now,
extracting metadata can become the slowest moment of update checking
(with up to 10x longer running times possible worst-case).

To keep the update checking times comparable whether or not synchronizing
ACL and xattrs, ctime is now used for fast metadata updates checking. To
this end, ctime is now stored in the archives and updated at every scan.