Persistent shared data for props
Until now, all the props have been individual atoms, tied to a specific
file and not knowing any context in the larger archive scope.

This commit adds the option for props to create shared data and persist
it in the archive.

The first use case for shared props data is hashing and deduplication of
synchronized ACL values. The expectation is that the number of distinct
ACL values is much smaller compared to the total number of files and
directories. By storing all unique ACL values within shared props data
and using a hash to look up values from the shared data, the memory
and archive storage requirements are reduced considerably (by up to
several times), with the trade-off of a tiny processing overhead.