Enable ctime updates checking in Windows
If correct ctime is available in Windows (it is not by default; most
likely the creation time is returned as ctime) then take advantage of
ctime-based updates checking.

The performance benefits are not known, but are likely to be modest at
best. The real benefit lies in detecting metadata-only updates, like a
change in ACL (on NTFS).