Get correct ctime in Windows
Standard Win32 API functions do not provide the status change time
(ctime) of files and directories. A common substitute is to provide
creation time, which is actually btime (and not available in the
traditional POSIX stat interface) and is not usable as ctime.

On some filesystems, it is possible to get the real ctime using Windows
syscalls via the Native API. Using Native API is a bit tricky, though.
Native API header files (for non-driver development?) are not available,
so relevant type definitions and function prototypes must be included in
your own code. Further, the functions are not available for build-time
linking and must be dynamically linked at runtime.
This approach also ensures compatibility with other compilers, like Mingw.

This patch borrows code from libuv to get the definitions and prototypes
and do the runtime linking.

If the Native API function NtQueryInformationFile is available then this
patch will use it, making correct ctime available in OCaml code, else it
will fall back to Win32 API.
Alternatively, NtQueryInformationByName could be used, with theoretical
performance gains, but this patch makes no attempt to use this function.

For file times, the previous code using Win32 API discarded second
fractions (not sure if this has been intentional). This patch does not
change that and uses the same time conversion formulas as is.