Load previous archive format (for upgrades)
Just like RPC/API compatibility with versions <= 2.51.4, it is possible
to load archives written by those versions. This is done only once, if
archives of the new format are not found. Archives in pervious format
are automatically deleted when the new archives are successfully written.

This code is intended to be temporary as its only purpose is to make the
upgrade smoother. It is not required for communication between different
versions.