Improve optimistic archive loading procedure
The optimistic archive loading can be described in simple terms as "when
there's no crash recovery needed and the archive file on disk has not
changed then use whatever is currently cached in memory, otherwise
re-load from the disk". The client and the server verify the optimistic
loading by checking that the archive checksums match. Should the
cheksums not match then the archive is loaded by the non-optimistic
procedure (just load the archive from disk, ignore anything cached in
memory).

There is a flaw in the checksum verification step, which combined with
the fact that both the optimistic and non-optimistic procedures have the
capability of loading from the disk, causes non-optimal archive loading
in some scenarios.

While the flaw also has the potential to cause more serious mismatch of
archive state between client and server (partially fixed by this patch),
this hasn't been an issue in practice because of careful coding elsewhere,
which has provided nearly a guarantee of archive consistency between
client and server.

The flaw lies in the fact that archive checksum is available only for
the archive that was stored on disk. In other words, the verification of
checksums between client and server can be done only if both have
freshly loaded the archive from the disk. Additionally, the optimistic
loading procedure is always tried first; only once the checksums didn't
match, it's followed by the non-optimistic archive loading. This has the
unfortunate side-effect of worst-case loading the archive from the disk
twice.

Perhaps to avoid the worst side-effects of this flaw, there is currently
a bypass to verifying checksums when both the client and server have the
archive cached in memory. This will usually work and not be a problem
because for a local sync, or when connecting with the server over ssh,
the client and server processes have the same lifecycle length and share
the same state (save for bugs).

When using the socket server, the server process can greatly outlive the
client process and several separate client processes can connect to the
server process over its lifetime.

When a new client connects to socket server, without this patch, the
optimistic archive loading will fail because the client will not have
cached the archive in memory, causing it to be loaded from the disk. The
server, on the other hand, does have the archive cached but can no
longer provide a checksum for it. Thus, both client and server now have
to go through the non-optimistic loading procedure, forcing the server
to unnecesasrily load the archive from disk and the client to again load
the archive from disk, even though it has just done that.

This patch caches the archive checksum so that it is available for the
optimistic archive loading procedure immediately after the archive has
been loaded from disk or after it was stored to disk, for as long as the
in-memory archive has not been changed. This is sufficient to prevent
the non-optimal archive loading with a socket server for normal usage.

Additionally, this patch re-calculates the archive checksum for the
in-memory archive just after the updates have been collected (as some
archive changes are done directly without proceeding through sync).
Calculating the checksum takes some time, of course, but is still an
order of magnitude faster than unnecessarily re-loading the archive from
disk. This change (1) avoids the non-optimal archive loading for the
next scan/new client when the previous sync with socket server was not
completed by previous client after scanning for updates; and (2)
provides some consistency verification for local/ssh sync when
rescanning the updates without running the previous sync after scanning
for updates.